requirements.txt
aiohttp==3.6.2
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.3.3
idna==2.9
idna-ssl==1.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pkg-resources==0.0.0
pylint==2.5.2
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
